Issue #1996: merge fasthttp-mysql and fasthttp-postgresql into fasthttp folder
