+++ .travis.yml
@@ -56,8 +56,7 @@ env:
     - "TESTDIR=Erlang/misultin"
     - "TESTDIR=Go/beego"
     - "TESTDIR=Go/falcore"
-    - "TESTDIR=Go/fasthttp-mysql"
-    - "TESTDIR=Go/fasthttp-postgresql"
+    - "TESTDIR=Go/fasthttp"
     - "TESTDIR=Go/gin"
     - "TESTDIR=Go/goji"
     - "TESTDIR=Go/go-std-mongodb"
