--- frameworks/Go/fasthttp-mysql/setup.bat
@@ -1,3 +1,0 @@
-set GOPATH=C:\FrameworkBenchmarks\Go\fasthttp-mysql
-go build -o server
-.\server
