--- frameworks/Go/fasthttp-mysql/source_code
@@ -1,4 +1,0 @@
-./fasthttp-mysql/server.go
-./fasthttp-mysql/src/
-./fasthttp-mysql/src/templates/
-./fasthttp-mysql/src/templates/fortune.qtpl
