--- frameworks/Go/fasthttp-postgresql/setup.bat
@@ -1,3 +1,0 @@
-set GOPATH=C:\FrameworkBenchmarks\Go\fasthttp-postgresql
-go build -o server
-.\server
