--- frameworks/Go/fasthttp-postgresql/source_code
@@ -1,4 +1,0 @@
-./fasthttp-postgresql/server.go
-./fasthttp-postgresql/src/
-./fasthttp-postgresql/src/templates/
-./fasthttp-postgresql/src/templates/fortune.qtpl
