+++ frameworks/Go/fasthttp/setup.bat
@@ -1,0 +1,3 @@
+set GOPATH=C:\FrameworkBenchmarks\Go\fasthttp
+go build server-mysql
+.\server-mysql
