+++ frameworks/Go/fasthttp/source_code
@@ -1,0 +1,8 @@
+./fasthttp/server.go
+./fasthttp/src/
+./fasthttp/src/common/common.go
+./fasthttp/src/server-mysql/server.go
+./fasthttp/src/server-postgresql/server.go
+./fasthttp/src/templates/
+./fasthttp/src/templates/aux.go
+./fasthttp/src/templates/fortune.qtpl
