+++ frameworks/Go/fasthttp/src/server-mysql/server.go
@@ -1,38 +1,21 @@
-//go:generate qtc -dir=src/templates
 package main
 
 import (
 	"database/sql"
-	"encoding/json"
 	"flag"
 	"log"
-	"math/rand"
-	"net"
-	"os"
-	"os/exec"
 	"runtime"
 	"sort"
-	"sync"
 
 	_ "github.com/go-sql-driver/mysql"
 	"github.com/valyala/fasthttp"
-	"github.com/valyala/fasthttp/reuseport"
 
+	"common"
 	"templates"
 )
 
-type JSONResponse struct {
-	Message string `json:"message"`
-}
-
-type World struct {
-	Id           uint16 `json:"id"`
-	RandomNumber uint16 `json:"randomNumber"`
-}
-
 const (
 	connectionString   = "benchmarkdbuser:benchmarkdbpass@tcp(localhost:3306)/hello_world"
-	worldRowCount      = 10000
 	maxConnectionCount = 40
 )
 
@@ -42,12 +25,6 @@ var (
 	fortuneSelectStmt *sql.Stmt
 
 	db *sql.DB
-)
-
-var (
-	listenAddr = flag.String("listenAddr", ":8080", "Address to listen to")
-	prefork    = flag.Bool("prefork", false, "use prefork")
-	child      = flag.Bool("child", false, "is child proc")
 )
 
 func main() {
@@ -62,7 +39,7 @@ func main() {
 	}
 
 	dbConnCount := maxConnectionCount
-	if *prefork {
+	if *common.Prefork {
 		dbConnCount = (dbConnCount + runtime.NumCPU() - 1) / runtime.NumCPU()
 	}
 	db.SetMaxIdleConns(dbConnCount)
@@ -76,7 +53,7 @@ func main() {
 		Handler: mainHandler,
 		Name:    "go",
 	}
-	ln := getListener()
+	ln := common.GetListener()
 	if err = s.Serve(ln); err != nil {
 		log.Fatalf("Error when serving incoming connections: %s", err)
 	}
@@ -86,9 +63,9 @@ func mainHandler(ctx *fasthttp.RequestCtx) {
 	path := ctx.Path()
 	switch string(path) {
 	case "/plaintext":
-		plaintextHandler(ctx)
+		common.PlaintextHandler(ctx)
 	case "/json":
-		jsonHandler(ctx)
+		common.JSONHandler(ctx)
 	case "/db":
 		dbHandler(ctx)
 	case "/queries":
@@ -102,40 +79,21 @@ func mainHandler(ctx *fasthttp.RequestCtx) {
 	}
 }
 
-// Test 1: JSON serialization
-func jsonHandler(ctx *fasthttp.RequestCtx) {
-	r := jsonResponsePool.Get().(*JSONResponse)
-	r.Message = "Hello, World!"
-	jsonMarshal(ctx, r)
-	jsonResponsePool.Put(r)
-}
-
-var jsonResponsePool = &sync.Pool{
-	New: func() interface{} {
-		return &JSONResponse{}
-	},
-}
-
-// Test 2: Single database query
 func dbHandler(ctx *fasthttp.RequestCtx) {
-	var w World
+	var w common.World
 	fetchRandomWorld(&w)
-	jsonMarshal(ctx, &w)
+	common.JSONMarshal(ctx, &w)
 }
 
-// Test 3: Multiple database queries
 func queriesHandler(ctx *fasthttp.RequestCtx) {
-	n := getQueriesCount(ctx)
-
-	worlds := make([]World, n)
+	n := common.GetQueriesCount(ctx)
+	worlds := make([]common.World, n)
 	for i := 0; i < n; i++ {
 		fetchRandomWorld(&worlds[i])
 	}
-
-	jsonMarshal(ctx, worlds)
+	common.JSONMarshal(ctx, worlds)
 }
 
-// Test 4: Fortunes
 func fortuneHandler(ctx *fasthttp.RequestCtx) {
 	rows, err := fortuneSelectStmt.Query()
 	if err != nil {
@@ -153,25 +111,24 @@ func fortuneHandler(ctx *fasthttp.RequestCtx) {
 	rows.Close()
 	fortunes = append(fortunes, templates.Fortune{Message: "Additional fortune added at request time."})
 
-	sort.Sort(FortunesByMessage(fortunes))
+	sort.Sort(common.FortunesByMessage(fortunes))
 
 	ctx.SetContentType("text/html; charset=utf-8")
 	templates.WriteFortunePage(ctx, fortunes)
 }
 
-// Test 5: Database updates
 func updateHandler(ctx *fasthttp.RequestCtx) {
-	n := getQueriesCount(ctx)
+	n := common.GetQueriesCount(ctx)
 
-	worlds := make([]World, n)
+	worlds := make([]common.World, n)
 	for i := 0; i < n; i++ {
 		w := &worlds[i]
 		fetchRandomWorld(w)
-		w.RandomNumber = uint16(randomWorldNum())
+		w.RandomNumber = int32(common.RandomWorldNum())
 	}
 
 	// sorting is required for insert deadlock prevention.
-	sort.Sort(WorldsByID(worlds))
+	sort.Sort(common.WorldsByID(worlds))
 	txn, err := db.Begin()
 	if err != nil {
 		log.Fatalf("Error starting transaction: %s", err)
@@ -187,54 +144,15 @@ func updateHandler(ctx *fasthttp.RequestCtx) {
 		log.Fatalf("Error when commiting world rows: %s", err)
 	}
 
-	jsonMarshal(ctx, worlds)
+	common.JSONMarshal(ctx, worlds)
 }
 
-// Test 6: Plaintext
-func plaintextHandler(ctx *fasthttp.RequestCtx) {
-	ctx.SetContentType("text/plain")
-	ctx.WriteString("Hello, World!")
-}
-
-func jsonMarshal(ctx *fasthttp.RequestCtx, v interface{}) {
-	ctx.SetContentType("application/json")
-	if err := json.NewEncoder(ctx).Encode(v); err != nil {
-		log.Fatalf("error in json.Encoder.Encode: %s", err)
-	}
-}
-
-func fetchRandomWorld(w *World) {
-	n := randomWorldNum()
+func fetchRandomWorld(w *common.World) {
+	n := common.RandomWorldNum()
 	if err := worldSelectStmt.QueryRow(n).Scan(&w.Id, &w.RandomNumber); err != nil {
 		log.Fatalf("Error scanning world row: %s", err)
 	}
 }
-
-func randomWorldNum() int {
-	return rand.Intn(worldRowCount) + 1
-}
-
-func getQueriesCount(ctx *fasthttp.RequestCtx) int {
-	n := ctx.QueryArgs().GetUintOrZero("queries")
-	if n < 1 {
-		n = 1
-	} else if n > 500 {
-		n = 500
-	}
-	return n
-}
-
-type FortunesByMessage []templates.Fortune
-
-func (s FortunesByMessage) Len() int           { return len(s) }
-func (s FortunesByMessage) Swap(i, j int)      { s[i], s[j] = s[j], s[i] }
-func (s FortunesByMessage) Less(i, j int) bool { return s[i].Message < s[j].Message }
-
-type WorldsByID []World
-
-func (w WorldsByID) Len() int           { return len(w) }
-func (w WorldsByID) Swap(i, j int)      { w[i], w[j] = w[j], w[i] }
-func (w WorldsByID) Less(i, j int) bool { return w[i].Id < w[j].Id }
 
 func mustPrepare(db *sql.DB, query string) *sql.Stmt {
 	stmt, err := db.Prepare(query)
@@ -242,41 +160,4 @@ func mustPrepare(db *sql.DB, query string) *sql.Stmt {
 		log.Fatalf("Error when preparing statement %q: %s", query, err)
 	}
 	return stmt
-}
-
-func getListener() net.Listener {
-	if !*prefork {
-		runtime.GOMAXPROCS(runtime.NumCPU())
-		ln, err := net.Listen("tcp4", *listenAddr)
-		if err != nil {
-			log.Fatal(err)
-		}
-		return ln
-	}
-
-	if !*child {
-		children := make([]*exec.Cmd, runtime.NumCPU())
-		for i := range children {
-			children[i] = exec.Command(os.Args[0], "-prefork", "-child")
-			children[i].Stdout = os.Stdout
-			children[i].Stderr = os.Stderr
-			if err := children[i].Start(); err != nil {
-				log.Fatal(err)
-			}
-		}
-		for _, ch := range children {
-			if err := ch.Wait(); err != nil {
-				log.Print(err)
-			}
-		}
-		os.Exit(0)
-		panic("unreachable")
-	}
-
-	runtime.GOMAXPROCS(1)
-	ln, err := reuseport.Listen("tcp4", *listenAddr)
-	if err != nil {
-		log.Fatal(err)
-	}
-	return ln
 }
