+++ frameworks/Go/fasthttp/src/templates/aux.go
@@ -1,0 +1,3 @@
+package templates
+
+//go:generate qtc
