+++ frameworks/Go/fasthttp/src/templates/fortune.qtpl
